[TACHYON-602] Login Mechanism: Add TACHYON_LOGIN_MODULE - address comments
